Overhaul feed integrity, live safety, latency, hot path, and PnL accounting
Implements the full review fix plan (verified against official Lighter and
Binance docs), plus the shadow->dry-run rename and step-clock vol sampling.
118 unit tests pass; validated end-to-end in dry-run against live feeds.

Feed integrity + live safety:
- Binance depth sync rewritten to the official USDS-M futures rules as a
  unit-tested state machine (strict u<lastUpdateId drop, bracketing first
  event, pu continuity, in-session re-snapshot). An empty/stale buffer no
  longer error-loops the connection on quiet symbols.
- Lighter book gap detection via begin_nonce/nonce continuity (offset is
  documented non-contiguous) with a forced fresh snapshot on gaps; ticker
  BBO now feeds a book-sanity cross-check instead of being discarded.
- Dead-man switch: stale market data (safety.md_deadman_sec) parks trading
  and cancels-all once per pause; the ladder is also pulled when capital
  reaches zero or the position feed is lost.
- Live startup sends update-leverage (imf=10000/leverage, margin_mode) so
  the venue matches the sizing math; config parse errors abort startup and
  live validates leverage/levels/sizing; test_sign now signs an offline
  update-leverage tx to verify signer FFI arity.
- REST position fetch demoted to a fallback so it cannot clobber fresh
  account-WS positions and spuriously re-arm the post-fill hold.

Post-fill latency:
- account_orders WS resolves client->exchange ids in ~100ms (mapping-only
  events; full reconcile stays REST-snapshot-driven).
- Own fills from account_all clear/shrink the exact slot immediately.
- The 10s full-ladder freeze is now a configurable 3s hold that blocks only
  Modify/Cancel; Creates keep flowing so filled levels re-quote instantly.

Hot path:
- One typed JSON parse per WS message (was Value parse + deep clone +
  re-parse); price levels parse to f64 at deserialize time; reusable level
  scratch buffers; SmallVec quote ladder with in-place transforms;
  tracked-ids published on change only; non-blocking logging; market-data
  task on a dedicated thread with optional core pinning
  (performance.pin_hot_core); Release/Acquire commit ordering on shared
  atomics.

Vol calibration:
- Vol/OBI windows now sample on the vol_obi.step_ns clock instead of every
  raw book message (dry-run measured ~20 updates/s vs the assumed 10/s,
  which underscaled volatility); alpha z-score still refreshes per tick.

PnL + housekeeping:
- Fill attribution by account id: fills without a client id (liquidations,
  manual) are booked to position accounting and a separate unattributed
  bucket instead of dropped.
- Pending-fill replay after a position-snapshot re-seed is PnL-only
  (sync-generation guard) - no more double-applied position/fees.
- MTM marks against the Lighter mid (Binance fallback) with a mid_stale
  flag instead of silently using an old price; live-state persistence is
  wired so realized PnL survives restarts (market_id u8->u32 fixed).
- PnL actor exits on channel close; bounded id sets; trade CSV can no
  longer block startup (first-line header check, rename-aside on legacy/
  corrupt) or duplicate rows after partial writes; summary writes are
  batched to the flush tick.
- Windows builds via cfg-gated instance lock (share-mode exclusive open);
  position_value_threshold_usd dust gate implemented; reduce-only exits
  sized to inventory; websocket.ping_interval wired; client-id 0 sentinel
  guarded; rejected batches count against the rate window; bookTicker idle
  timeout raised for BBO-change-driven quiet periods; dry-run synthesizes
  binds so the modify/cancel flow is exercised.

Rename: Mode::Shadow -> Mode::DryRun, --dry-run flag (--shadow kept as a
deprecated alias), DRY-RUN log labels, Dockerfile/README/DOCKER.md updated.

+++ src/account/persistence.rs
@@ -56,9 +56,10 @@ pub struct LiveState {
     /// Lighter account index this state belongs to.
     #[serde(default)]
     pub account_index: i64,
-    /// Market id this state belongs to.
+    /// Market id this state belongs to (u32 to match `MarketConfig::market_id` — the old
+    /// `u8` silently truncated ids > 255).
     #[serde(default)]
-    pub market_id: u8,
+    pub market_id: u32,
     /// Local signed position-size estimate (`_live_fill_position_size`).
     #[serde(default)]
     pub position_size_est: f64,
